Implement profile section
Now the content of the profile section is being stored in
data/profile.toml.

The amount of years showcase are being stored in data/years.toml.

The social newtork links are being rendered based on the file
data/social.toml.

And if the hugo site has a configuration called "copyright" it will get
its content dispalyed in the site's footer.